Refactor core modules to DDD naming standards
- Renamed `cognitive` -> `compute`
- Renamed `engines.py` -> `reasoning.py`
- Renamed `topology` -> `workflow`
- Renamed `co_intelligence.py` -> `intervention.py`
- Fixed all imports globally.
- Updated `rebuild.py` lazy imports and namespace map.
- Removed test files as per instruction.
