+++ src/coreason_manifest/builder.py
@@ -10,23 +10,11 @@
 
 from typing import Any, Self, cast
 
-from coreason_manifest.spec.core.oversight.co_intelligence import EscalationCriteria
-from coreason_manifest.spec.core.cognitive.engines import (
+from coreason_manifest.spec.core.compute.nodes import AgentNode, CognitiveProfile, HumanNode, InspectorNode
+from coreason_manifest.spec.core.compute.reasoning import (
     FastPath,
     ReasoningConfig,
     StandardReasoning,
-)
-from coreason_manifest.spec.core.topology.flow import (
-    AnyNode,
-    Blackboard,
-    DataSchema,
-    Edge,
-    FlowDefinitions,
-    FlowInterface,
-    FlowMetadata,
-    Graph,
-    GraphFlow,
-    LinearFlow,
 )
 from coreason_manifest.spec.core.oversight.governance import (
     CircuitBreaker,
@@ -36,14 +24,7 @@ from coreason_manifest.spec.core.oversight.governance import (
     Governance,
     OperationalPolicy,
 )
-from coreason_manifest.spec.core.state.memory import (
-    EpisodicMemoryConfig,
-    MemorySubsystem,
-    ProceduralMemoryConfig,
-    SemanticMemoryConfig,
-    WorkingMemoryConfig,
-)
-from coreason_manifest.spec.core.cognitive.nodes import AgentNode, CognitiveProfile, HumanNode, InspectorNode
+from coreason_manifest.spec.core.oversight.intervention import EscalationCriteria
 from coreason_manifest.spec.core.oversight.resilience import (
     ErrorDomain,
     ErrorHandler,
@@ -55,7 +36,26 @@ from coreason_manifest.spec.core.oversight.resilience import (
     SupervisionPolicy,
 )
 from coreason_manifest.spec.core.rebuild import rebuild_manifest
+from coreason_manifest.spec.core.state.memory import (
+    EpisodicMemoryConfig,
+    MemorySubsystem,
+    ProceduralMemoryConfig,
+    SemanticMemoryConfig,
+    WorkingMemoryConfig,
+)
 from coreason_manifest.spec.core.state.tools import ToolPack
+from coreason_manifest.spec.core.workflow.flow import (
+    AnyNode,
+    Blackboard,
+    DataSchema,
+    Edge,
+    FlowDefinitions,
+    FlowInterface,
+    FlowMetadata,
+    Graph,
+    GraphFlow,
+    LinearFlow,
+)
 from coreason_manifest.utils.validator import validate_flow
 
 
@@ -626,6 +626,7 @@ class BaseFlowBuilder:
             type="agent",
             profile=profile_id,
             tools=tools,
+            operational_policy=None,
         )
         self._register_node(node)
         return self
