+++ src/coreason_manifest/spec/core/compute/reasoning.py
