+++ src/coreason_manifest/spec/core/oversight/intervention.py
@@ -60,3 +60,5 @@ class CoIntelligencePolicy(CoreasonModel):
         None,
         description="Default escalation strategy for global interventions.",
     )
+
+__all__ = ["CoIntelligencePolicy", "EscalationCriteria", "InterventionMode"]
