+++ src/coreason_manifest/utils/openai_adapter.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from coreason_manifest.spec.core.cognitive.nodes import AgentNode
+from coreason_manifest.spec.core.compute.nodes import AgentNode
 from coreason_manifest.spec.core.state.tools import ToolPack
 from coreason_manifest.spec.interop.adapter_config import AdapterConfig
 from coreason_manifest.spec.interop.compliance import RemediationAction
