+++ src/coreason_manifest/utils/validator.py
@@ -5,15 +5,7 @@ from typing import Any
 
 from pydantic import BaseModel
 
-from coreason_manifest.spec.core.topology.flow import (
-    AnyNode,
-    FlowDefinitions,
-    Graph,
-    GraphFlow,
-    LinearFlow,
-)
-from coreason_manifest.spec.core.oversight.governance import Governance
-from coreason_manifest.spec.core.cognitive.nodes import (
+from coreason_manifest.spec.core.compute.nodes import (
     AgentNode,
     CognitiveProfile,
     EmergenceInspectorNode,
@@ -21,16 +13,23 @@ from coreason_manifest.spec.core.cognitive.nodes import (
     SwarmNode,
     SwitchNode,
 )
+from coreason_manifest.spec.core.oversight.governance import Governance
 from coreason_manifest.spec.core.oversight.resilience import (
     EscalationStrategy,
     FallbackStrategy,
     ReflexionStrategy,
     ResilienceStrategy,
 )
-from coreason_manifest.spec.core.state.tools import ToolCapability, ToolPack
 from coreason_manifest.spec.core.primitives.types import RiskLevel
+from coreason_manifest.spec.core.state.tools import ToolCapability, ToolPack
+from coreason_manifest.spec.core.workflow.flow import (
+    AnyNode,
+    FlowDefinitions,
+    Graph,
+    GraphFlow,
+    LinearFlow,
+)
 from coreason_manifest.spec.interop.compliance import ComplianceReport, ErrorCatalog, RemediationAction
-from coreason_manifest.spec.interop.exceptions import ManifestError, ManifestErrorCode
 from coreason_manifest.utils.topology import get_strongly_connected_components, get_unified_topology
